// src/tests/userTickets/createUserTicket/createUserTicket.generated.ts
/* eslint-disable */
/* @ts-nocheck */
/* prettier-ignore */
/* This file is automatically generated using `npm run graphql:types` */
import type * as Types from '../../../generated/types';

import type { JsonObject } from "type-fest";
import gql from 'graphql-tag';
export type CreateUserTicketMutationVariables = Types.Exact<{
  ticketId: Types.Scalars['String']['input'];
}>;


export type CreateUserTicketMutation = { __typename?: 'Mutation', createUserTicket: { __typename?: 'UserTicket', id: string, approvalStatus: Types.TicketApprovalStatus, paymentStatus: Types.TicketPaymentStatus, redemptionStatus: Types.TicketRedemptionStatus, status: Types.TicketStatus } };


export const CreateUserTicket = gql`
    mutation CreateUserTicket($ticketId: String!) {
  createUserTicket(ticketId: $ticketId) {
    id
    approvalStatus
    paymentStatus
    redemptionStatus
    status
  }
}
    `;